requirements.txt
streamlit==0.79.0
st-annotated-text
